% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/work_test_norm_global.R
\name{adaptive_analysis_norm_global}
\alias{adaptive_analysis_norm_global}
\title{Analyze data according to a globally efficient adaptive design.}
\usage{
adaptive_analysis_norm_global(initial_test = 0, times = 0, stats = 0,
  costs = 0, final_analysis = TRUE, estimate = TRUE, ci_coef = 0.95,
  tol_est = 1e-08, input_check = TRUE)
}
\arguments{
\item{initial_test}{Designate the initial working test generated by \code{work_test_norm_global} function.}

\item{times}{The sequence of times (sample size or information level) at which analyses were conducted.}

\item{stats}{The sequence of test statistics.}

\item{costs}{The sequence of loss required to construct working tests. Specification is optional. Partial specification is allowed, in which non-specification may be represented by \code{0}.}

\item{final_analysis}{If \code{TRUE}, the result input will be regarded as complete (no more data will be obtained) and the significance level will be exhausted. If \code{FALSE}, the current analysis will be regarded as an interim analysis and the significance level will be preserved.}

\item{estimate}{If \code{TRUE}, p-value, median unbiased estimator and upper and lower confidence limits will be calculated.}

\item{ci_coef}{The confidence coefficient. Default is 0.95.}

\item{tol_est}{The precision of the calculated results.}

\item{input_check}{Indicate whether or not the arguments input by user contain invalid values.}
}
\value{
It returns whether or not the result was statistically significant, a p-value and an exact confidence limits.
}
\description{
\code{adaptive_analysis_norm_global} performs an globally efficient adaptive test, 
a Frequentist adaptive test with the specified significance level
with full flexibility.
Normality with known variance is assumed for the test statistic
(more accurately, the test statistic is assumed to follow Brownian motion.)
Null hypothesis is fixed at 0 without loss of generality.
No procedure to calculate p-value or confidence intervals is employed.
}
\examples{
# Loss by erroneous rejection of the null hypothesis
# (These values can be calculated if its input is omitted in the functions below)
cost0 <- c(1683.458, 1555.020, 1545.278, 1528.397, 1471.727)

# Construct an initial working test
init_work_test <- work_test_norm_global(min_effect_size = -log(0.65), cost_type_1_err=cost0[1])
with(init_work_test, plot(par$U_0, char$boundary, xlim=range(0, par$U_0),
  ylim=range(0, char$boundary[-1]), pch=16, cex=0.5) )

# Final interim analysis
interim_analysis_4 <- adaptive_analysis_norm_global(
  initial_test = init_work_test,
  times = c(5.67, 9.18, 14.71, 20.02),
  stats = c(3.40, 4.35, 7.75, 11.11),
  costs = cost0[-1],
  final_analysis = FALSE,
  estimate = FALSE
  )
print( with(interim_analysis_4, data.frame(analysis=0:par$analyses, time=par$times,
  cost=char$cost0, stat=par$stats, boundary=char$boundary, pr_cond_err=char$cond_type_I_err,
  reject_H0=char$rej_H0)) )

# Sample size calculation
sample_size_norm_global(
  initial_test = init_work_test,
  effect_size = 11.1110 / 20.02, # needs not be MLE
  time = 20.02,
  target_power = 0.75,
  sample_size = TRUE
  )

# Final analysis
final_analysis <- adaptive_analysis_norm_global(
  initial_test = init_work_test,
  times = c(5.67, 9.18, 14.71, 20.02, 24.86),
  stats = c(3.40, 4.35, 7.75, 11.11, 14.84),
  costs = cost0[-1],
  final_analysis = TRUE,
  estimate = FALSE
  )
print( with(final_analysis, data.frame(analysis=0:par$analyses, time=par$times,
  cost=char$cost0, stat=par$stats, boundary=char$boundary, pr_cond_err=char$cond_type_I_err,
  reject_H0=char$rej_H0)) )
}
\references{
Kashiwabara, K., Matsuyama, Y. An efficient adaptive design approximating fixed sample size designs. In preparation.
}
\seealso{
\code{\link{work_test_norm_global}} and \code{\link{sample_size_norm_global}}.
}
